Add promote command.
Promoting a file completes the editing workflow started by the version
command, replacing an archive file with a new version from a
workspace while allowing for the file extension to change.

+++ lib/ImageArchive/Archive.rakumod
@@ -6,7 +6,7 @@ use ImageArchive::Config;
 use ImageArchive::Database;
 use ImageArchive::Util;
 
-sub deleteAlts(IO::Path $file) {
+sub deleteAlts(IO::Path $file) is export {
     my $cacheRoot = getPath('cache');
 
     my $relativePath = relativePath($file.Str);
@@ -15,6 +15,7 @@ sub deleteAlts(IO::Path $file) {
 
     for readConfig('alt_sizes').split(' ') -> $size {
         my $target = $cacheRoot.add("$size/$relativePath").extension($thumbnailExtension);
+        next unless $target ~~ :f;
         $target.IO.unlink;
         deleteEmptyFolders($target.parent);
     }
